[Bug #19728] Auto-generate unicode property docs
https://bugs.ruby-lang.org/issues/19728

This PR should probably wait for #7923 to be completed. See #7923 for detailed discussion.

The main open question to me is whether there is a better way to categorize the properties, or to name or order the sections? The current choices are based on https://github.com/k-takata/Onigmo/blob/master/doc/UnicodeProps.txt